# Improved performance by fetching column names once