refs https://github.com/TryGhost/node-sqlite3/pull/1471/commits/ad569c00d642650347555ddea38e126a04fcf788

- this commit optimizes the library by only fetching column names
  once whilst preparing the data to return
- we save a lot of blocking calls to the sqlite3 lib and this improves `.each` and `.all`, which would ordinarily do this once per row
- `benchmark/select.js` statistics go from `db.each: 92.393ms, db.all:
  101.076ms` to `db.each: 83.254ms, db.all: 93.211ms`, which is a ~9-10%
  improvement